[#109] remove references to RupSetDiagnosticsPageGen
