+++ .github/workflows/gradle.yml
@@ -18,7 +18,7 @@ jobs:
     - name: Clone opensha
       uses: actions/checkout@v2
       with:
-        repository: GNS-Science/opensha
+        repository: opensha/opensha
         ref: master
         path: opensha
